revert: move grid_ back to original location in pooler – no need to make two changes in this one PR
