+++ docs/experimentalists/pooler/grid/index.md
@@ -25,7 +25,7 @@ This means that there are various combinations that these variables can form, th
 
 ```python
 
-from autora.experimentalist.grid_ import grid_pool
+from autora.experimentalist.pooler.grid import grid_pool
 from autora.variable import Variable
 
 iv_1 = Variable(allowed_values=[1, 2, 3])
